% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{maf_inverse}
\alias{maf_inverse}
\title{base variable u -> theta (inverts maf_forward), dimension by dimension}
\usage{
maf_inverse(net, u, x)
}
\description{
base variable u -> theta (inverts maf_forward), dimension by dimension
}
\keyword{internal}
