% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{maf_module}
\alias{maf_module}
\title{The full MAF: a stack of MADE transforms with order reversal in between}
\usage{
maf_module(dim_x, dim_theta, n_transforms, hidden, embedding = NULL)
}
\description{
The full MAF: a stack of MADE transforms with order reversal in between
}
\keyword{internal}
