% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf.R
\name{nsf_inverse}
\alias{nsf_inverse}
\title{base variable u -> theta, inverting each spline dimension-by-dimension}
\usage{
nsf_inverse(net, u, x)
}
\description{
base variable u -> theta, inverting each spline dimension-by-dimension
}
\keyword{internal}
