% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf.R
\name{nsf_made_module}
\alias{nsf_made_module}
\title{MADE block emitting 3K - 1 spline parameters per dimension}
\usage{
nsf_made_module(dim_x, dim_theta, hidden, n_bins)
}
\description{
MADE block emitting 3K - 1 spline parameters per dimension
}
\keyword{internal}
