% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf.R
\name{nsf_module}
\alias{nsf_module}
\title{The full NSF: stacked spline-autoregressive transforms with order reversal}
\usage{
nsf_module(
  dim_x,
  dim_theta,
  n_transforms,
  hidden,
  n_bins,
  tail_bound,
  embedding = NULL
)
}
\description{
The full NSF: stacked spline-autoregressive transforms with order reversal
}
\keyword{internal}
